Remove censysCertificiates and censysIpv4; move santizeStringFields() to helper; move interface IpToDomainsMap to shodan.ts.
